Setting the precip data type to float32.

+++ swafi/precip.py
@@ -293,7 +293,8 @@ class Precipitation:
 
             with open(original_file, 'rb') as f_in:
                 data = pickle.load(f_in)
-                data[self.precip_var] = (data[self.precip_var] - mean) / std
+                data[self.precip_var] = ((data[self.precip_var] - mean) / std
+                                         ).astype('float32')
 
                 with open(tmp_filename, 'wb') as f_out:
                     pickle.dump(data, f_out)
@@ -321,7 +322,7 @@ class Precipitation:
 
             with open(original_file, 'rb') as f_in:
                 data = pickle.load(f_in)
-                data[self.precip_var] = data[self.precip_var] / q95
+                data[self.precip_var] = (data[self.precip_var] / q95).astype('float32')
 
                 with open(tmp_filename, 'wb') as f_out:
                     pickle.dump(data, f_out)
